+++ dev_scripts/qa.py
@@ -45,7 +45,7 @@ poetry run ./dev_scripts/qa.py {distro}-{version}
   - [ ] Create a new development environment with Poetry.
   - [ ] Build the container image and ensure the development environment uses
     the new image.
-  - [ ] Download the OCR language data using `./install/common/download-tessdata.py`
+  - [ ] Download the necessary assets using `poetry run assets install`
   - [ ] Run the Dangerzone tests.
   - [ ] Build and run the Dangerzone .exe
   - [ ] Test some QA scenarios (see [Scenarios](#Scenarios) below).
@@ -54,7 +54,7 @@ poetry run ./dev_scripts/qa.py {distro}-{version}
   - [ ] Create a new development environment with Poetry.
   - [ ] Build the container image and ensure the development environment uses
     the new image.
-  - [ ] Download the OCR language data using `./install/common/download-tessdata.py`
+  - [ ] Download the necessary assets using `poetry run assets install`
   - [ ] Run the Dangerzone tests.
   - [ ] Create and run an app bundle.
   - [ ] Test some QA scenarios (see [Scenarios](#Scenarios) below).
@@ -63,7 +63,7 @@ poetry run ./dev_scripts/qa.py {distro}-{version}
   - [ ] Create a new development environment with Poetry.
   - [ ] Build the container image and ensure the development environment uses
     the new image.
-  - [ ] Download the OCR language data using `./install/common/download-tessdata.py`
+  - [ ] Download the necessary assets using `poetry run assets install`
   - [ ] Run the Dangerzone tests.
   - [ ] Create and run an app bundle.
   - [ ] Test some QA scenarios (see [Scenarios](#Scenarios) below).
@@ -72,7 +72,7 @@ poetry run ./dev_scripts/qa.py {distro}-{version}
   - [ ] Create a new development environment with Poetry.
   - [ ] Build the container image and ensure the development environment uses
     the new image.
-  - [ ] Download the OCR language data using `./install/common/download-tessdata.py`
+  - [ ] Download the necessary assets using `poetry run assets install`
   - [ ] Run the Dangerzone tests.
   - [ ] Create a .deb package and install it system-wide.
   - [ ] Test some QA scenarios (see [Scenarios](#Scenarios) below).
@@ -81,7 +81,7 @@ poetry run ./dev_scripts/qa.py {distro}-{version}
   - [ ] Create a new development environment with Poetry.
   - [ ] Build the container image and ensure the development environment uses
     the new image.
-  - [ ] Download the OCR language data using `./install/common/download-tessdata.py`
+  - [ ] Download the necessary assets using `poetry run assets install`
   - [ ] Run the Dangerzone tests.
   - [ ] Create an .rpm package and install it system-wide.
   - [ ] Test some QA scenarios (see [Scenarios](#Scenarios) below).
@@ -292,10 +292,12 @@ Build the latest container:
 python3 ./install/common/build-image.py
 ```
 
-Download the OCR language data:
+Dangerzone depends on some assets that should be downloaded in order to run
+(think binaries and others resources). This can be done with the following
+command:
 
 ```sh
-python3 ./install/common/download-tessdata.py
+poetry run assets install
 ```
 
 Run from source tree:
@@ -355,10 +357,12 @@ Build the latest container:
 python3 ./install/common/build-image.py
 ```
 
-Download the OCR language data:
+Dangerzone depends on some assets that should be downloaded in order to run
+(think binaries and others resources). This can be done with the following
+command:
 
 ```sh
-python3 ./install/common/download-tessdata.py
+poetry run assets install
 ```
 
 Run from source tree:
@@ -419,10 +423,12 @@ Build the dangerzone container image:
 python3 .\install\common\build-image.py
 ```
 
-Download the OCR language data:
+Dangerzone depends on some assets that should be downloaded in order to run
+(think binaries and others resources). This can be done with the following
+command:
 
 ```sh
-python3 .\install\common\download-tessdata.py
+poetry run assets install
 ```
 
 After that you can launch dangerzone during development with:
@@ -758,9 +764,9 @@ class QABase(abc.ABC):
                 self.prompt("Does it pass?", choices=["y", "n"])
         logger.info("Successfully completed QA scenarios")
 
-    @task("Download Tesseract data", auto=True)
-    def download_tessdata(self):
-        self.run("python", str(Path("install", "common", "download-tessdata.py")))
+    @task("Download the necessary assets", auto=True)
+    def instal_assets(self):
+        self.run("poetry", "run", "assets", "install")
 
     @classmethod
     @abc.abstractmethod
@@ -862,7 +868,7 @@ class QAWindows(QABase):
         self.install_docker()
         self.install_poetry()
         self.build_image()
-        self.download_tessdata()
+        self.install_assets()
         self.run_tests()
         self.build_dangerzone_exe()
 
@@ -954,7 +960,7 @@ class QALinux(QABase):
     def start(self):
         self.build_dev_image()
         self.build_container_image()
-        self.download_tessdata()
+        self.install_assets()
         self.run_tests()
         self.build_package()
         self.build_qa_image()